Move cpus and disassemblers.
