+++ include/arviss/basic_rv32i_cpu.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "arviss/basic_mem.h"
-#include "arviss/rv32i_cpu.h"
+#include "arviss/rv32/cpus.h"
 
 namespace arviss
 {
